+++ backend/src/baserow/api/user/jwt.py
@@ -1,17 +1,68 @@
+from typing import Dict, Optional, Type
+
+from django.contrib.auth import get_user_model
+from django.contrib.auth.models import AbstractUser
+
+from rest_framework.request import Request
+from rest_framework_simplejwt.settings import api_settings
+from rest_framework_simplejwt.tokens import AccessToken, Token, TokenError
+
 from baserow.api.user.registries import user_data_registry
 
-from .serializers import UserSerializer
+
+def get_user_from_jwt_token(
+    token: str,
+    token_class: Optional[Type[Token]] = None,
+) -> AbstractUser:
+    """
+    Returns the user that is associated with the given JWT token.
+
+    :param token: The JWT token string
+    :param token_class: The token class that must be used to decode the token.
+    :param request: The request that is used to get the user data.
+    :return: The user that is associated with the token
+    :raises TokenError: If the token is invalid or if the user does not exist.
+    """
+
+    TokenClass = token_class or AccessToken
+    payload = TokenClass(token).token_backend.decode(token)
+    user_id = payload.get(api_settings.USER_ID_CLAIM)
+    User = get_user_model()
+    try:
+        return User.objects.get(pk=user_id)
+    except User.DoesNotExist:
+        # It could happen if the user was deleted after the token was issued.
+        raise TokenError("User does not exist.")
 
 
-def jwt_response_payload_handler(token, user=None, request=None, issued_at=None):
-    payload = {
-        "token": token,
-        "user": UserSerializer(user, context={"request": request}).data,
-    }
-
+def get_all_user_data_serialized(
+    user: AbstractUser, request: Optional[Request] = None
+) -> Dict:
     # Update the payload with the additional user data that must be added. The
     # `user_data_registry` contains instances that want to add additional information
     # to this payload.
-    payload.update(**user_data_registry.get_all_user_data(user, request))
+    from baserow.api.user.serializers import UserSerializer
 
-    return payload
+    return {
+        "user": UserSerializer(user, context={"request": request}).data,
+        **user_data_registry.get_all_user_data(user, request),
+    }
+
+
+def get_all_user_data_serialized_from_jwt_token(
+    token: str,
+    token_class: Optional[Type[Token]] = None,
+    request: Optional[Request] = None,
+) -> Dict:
+    """
+    Returns a dictionary with all the user data that is associated with the given JWT
+    token.
+
+    :param token: The JWT token string
+    :param token_class: The token class that must be used to decode the token.
+    :param request: The request that is used to get the user data.
+    :return: The user data that is associated with the token
+    """
+
+    user = get_user_from_jwt_token(token, token_class)
+    return get_all_user_data_serialized(user, request)