+++ backend/src/baserow/ws/auth.py
@@ -2,19 +2,14 @@ import uuid
 from urllib.parse import parse_qs
 
 from django.conf import settings
-from django.contrib.auth import get_user_model
 
 from channels.db import database_sync_to_async
 from channels.middleware import BaseMiddleware
 from rest_framework_simplejwt.exceptions import TokenError
 
-from baserow.api.authentication import get_user_from_jwt_token
-
 # Nosec disables spurious hardcoded password warning, this is not a password but instead
 # the value of the JWT token to be used when a user wants to connect anonymously.
 ANONYMOUS_USER_TOKEN = "anonymous"  # nosec
-
-User = get_user_model()
 
 
 @database_sync_to_async
@@ -29,6 +24,8 @@ def get_user(token):
     :rtype: User or None
     """
 
+    from baserow.api.user.jwt import get_user_from_jwt_token
+
     anonymous = token == ANONYMOUS_USER_TOKEN
     if anonymous:
         if settings.DISABLE_ANONYMOUS_PUBLIC_VIEW_WS_CONNECTIONS:
@@ -40,7 +37,7 @@ def get_user(token):
     else:
         try:
             return get_user_from_jwt_token(token)
-        except (TokenError, User.DoesNotExist):
+        except TokenError:
             return